Moved AbortableAsyncIterator to `index.ts`
Closes https://github.com/ollama/ollama-js/issues/135
