+++ src/index.ts
@@ -1,11 +1,10 @@
 import * as utils from './utils.js'
-import { AbortableAsyncIterator } from './utils.js'
 
 import fs, { createReadStream, promises } from 'fs'
 import { dirname, join, resolve } from 'path'
 import { createHash } from 'crypto'
 import { homedir } from 'os'
-import { Ollama as OllamaBrowser } from './browser.js'
+import { type ErrorResponse, Ollama as OllamaBrowser } from './browser.js'
 
 import type { CreateRequest, ProgressResponse } from './interfaces.js'
 
@@ -170,6 +169,41 @@ export class Ollama extends OllamaBrowser {
 }
 
 export default new Ollama()
+
+/**
+ * An AsyncIterator which can be aborted
+ */
+export class AbortableAsyncIterator<T extends object> {
+  private readonly abortController: AbortController
+  private readonly itr: AsyncGenerator<T | ErrorResponse>
+  private readonly doneCallback: () => void
+
+  constructor(abortController: AbortController, itr: AsyncGenerator<T | ErrorResponse>, doneCallback: () => void) {
+    this.abortController = abortController
+    this.itr = itr
+    this.doneCallback = doneCallback
+  }
+
+  abort() {
+    this.abortController.abort()
+  }
+
+  async *[Symbol.asyncIterator]() {
+    for await (const message of this.itr) {
+      if ('error' in message) {
+        throw new Error(message.error)
+      }
+      yield message
+      // message will be done in the case of chat and generate
+      // message will be success in the case of a progress response (pull, push, create)
+      if ((message as any).done || (message as any).status === 'success') {
+        this.doneCallback()
+        return
+      }
+    }
+    throw new Error('Did not receive done or success response in stream.')
+  }
+}
 
 // export all types from the main entry point so that packages importing types dont need to specify paths
 export * from './interfaces.js'
